fix dependencies and compilations warnings
This commit moves all dependencies that don't go into the final library
functions to `[dev-dependencies]` in `Cargo.toml`. Like this, we don't
polute users of our library with dependencies they don't need. This has
great benefits also in terms of compilation speed. I also went ahead and
updated our dependencies.

Also, I added a `.gitignore` and a `rust-toolchain.toml`. The second is
extremelly important since this crate only compiles with the nightly
toolchain.

Finally, this commit fixes some warnings the rustc was giving when
trying to compile the crate.